EE-518: Move contract_api::Error to its own module
